rearranges ML scripts into separate ml/ folder. adjusts references to ml_test to point to new location.
